+++ src/hotspot/share/compiler/compileBroker.cpp
@@ -934,7 +934,7 @@ void CompileBroker::init_compiler_threads() {
     // Create a name for our thread.
     int printed_len = os::snprintf(name_buffer, sizeof(name_buffer), "%s CompilerThread%d", _compilers[1]->name(), i);
     assert(printed_len > 0, "error occurs at os::snprintf");
-    assert(printed_len < sizeof(name_buffer), "name_buffer overflow");
+    assert((size_t)printed_len < sizeof(name_buffer), "name_buffer overflow");
     Handle thread_oop = create_thread_oop(name_buffer, CHECK);
     thread_handle = JNIHandles::make_global(thread_oop);
     JVMCI_ONLY(})
@@ -958,7 +958,7 @@ void CompileBroker::init_compiler_threads() {
     // Create a name for our thread.
     int printed_len = os::snprintf(name_buffer, sizeof(name_buffer), "C1 CompilerThread%d", i);
     assert(printed_len > 0, "error occurs at os::snprintf");
-    assert(printed_len < sizeof(name_buffer), "name_buffer overflow");
+    assert((size_t)printed_len < sizeof(name_buffer), "name_buffer overflow");
     Handle thread_oop = create_thread_oop(name_buffer, CHECK);
     jobject thread_handle = JNIHandles::make_global(thread_oop);
     _compiler1_objects[i] = thread_handle;
@@ -1024,7 +1024,7 @@ void CompileBroker::possibly_add_compiler_threads(JavaThread* THREAD) {
         char name_buffer[256];
         int printed_len = os::snprintf(name_buffer, sizeof(name_buffer), "%s CompilerThread%d", _compilers[1]->name(), i);
         assert(printed_len > 0, "error occurs at os::snprintf");
-        assert(printed_len < sizeof(name_buffer), "name_buffer overflow");
+        assert((size_t)printed_len < sizeof(name_buffer), "name_buffer overflow");
         Handle thread_oop;
         {
           // We have to give up the lock temporarily for the Java calls.
@@ -2608,7 +2608,7 @@ void CompileBroker::print_times(bool per_compiler, bool aggregate) {
       CompilerStatistics* stats = &_stats_per_level[tier-1];
       int printed_len = os::snprintf(tier_name, sizeof(tier_name), "Tier%d", tier);
       assert(printed_len > 0, "error occurs at os::snprintf");
-      assert(printed_len < sizeof(tier_name), "tier_name buf overflow");
+      assert((size_t)printed_len < sizeof(tier_name), "tier_name buf overflow");
       print_times(tier_name, stats);
     }
   }
