+++ src/hotspot/share/runtime/os.cpp
@@ -1275,7 +1275,7 @@ FILE* os::fopen(const char* path, const char* mode) {
   assert(strlen(mode) + 1 < sizeof(modified_mode), "mode chars plus one extra must fit in buffer");
   int printed_len = os::snprintf(modified_mode, sizeof(modified_mode), "%s" LINUX_ONLY("e") BSD_ONLY("e") WINDOWS_ONLY("N"), mode);
   assert(printed_len > 0, "error occurs at os::snprintf");
-  assert(printed_len < sizeof(modified_mode), "modified_mode buf overflow");
+  assert((size_t)printed_len < sizeof(modified_mode), "modified_mode buf overflow");
   FILE* file = ::fopen(path, modified_mode);
 
 #if !(defined LINUX || defined BSD || defined _WINDOWS)
